feat: securing profile & product views #7

+++ products/views.py
@@ -4,6 +4,7 @@ from django.contrib import messages
 from django.db.models import Q # This is for searching queries
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
 
 
 def all_products(request):
@@ -95,10 +96,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ 
     Add a product to the store 
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -119,10 +125,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ 
     Edit a product to the store 
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -143,10 +154,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ 
     Delete a product to the store 
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owner can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
